app/services/requirements.txt
acres==0.3.0
annotated-types==0.7.0
anyio==4.9.0
blinker==1.9.0
boto3==1.37.29
botocore==1.37.29
certifi==2025.1.31
charset-normalizer==3.4.1
ci-info==0.3.0
click==8.1.8
configobj==5.0.9
configparser==7.2.0
distro==1.9.0
etelemetry==0.3.1
filelock==3.18.0
Flask==3.1.0
h11==0.14.0
httpcore==1.0.7
httplib2==0.22.0
httpx==0.28.1
idna==3.10
isodate==0.6.1
itsdangerous==2.2.0
Jinja2==3.1.6
jiter==0.9.0
jmespath==1.0.1
looseversion==1.3.0
lxml==5.3.2
MarkupSafe==3.0.2
networkx==3.4.2
nibabel==5.3.2
nipype==1.10.0
numpy==2.2.4
openai==1.72.0
packaging==24.2
pandas==2.2.3
pathlib==1.0.1
prov==2.0.1
puremagic==1.28
pydantic==2.11.3
pydantic_core==2.33.1
pydot==3.0.4
PyMuPDF==1.25.5
pyparsing==3.2.3
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
pytz==2025.2
pyxnat==1.6.3
rdflib==6.3.2
requests==2.32.3
s3transfer==0.11.4
scipy==1.15.2
simplejson==3.20.1
six==1.17.0
sniffio==1.3.1
tqdm==4.67.1
traits==7.0.2
typing-inspection==0.4.0
typing_extensions==4.13.1
tzdata==2025.2
urllib3==2.3.0
Werkzeug==3.1.3
